requirements.txt
absl-py==1.4.0
addict==2.4.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
contourpy==1.1.0
cxxfilt==0.3.0
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
fire==0.5.0
flake8==6.1.0
fonttools==4.42.0
fqdn==1.5.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
llvmlite==0.40.1
lyft-dataset-sdk==0.0.8
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
MinkowskiEngine @ git+https://github.com/NVIDIA/MinkowskiEngine@02fc608bea4c0549b0a7b00ca1bf15dee4a0b228
mistune==3.0.1
mmcls==0.25.0
mmcv==1.7.0
mmcv-full==1.7.1
-e git+https://github.com/open-mmlab/mmdetection.git@e9cae2d0787cd5c2fc6165a6061f92fa09e48fb1#egg=mmdet
-e git+https://github.com/open-mmlab/mmdetection3d.git@47285b3f1e9dba358e98fcd12e523cfd0769c876#egg=mmdet3d
mmengine==0.8.4
-e git+https://github.com/open-mmlab/mmsegmentation.git@ed839828760a5f6193822e0bf3492b88ae6140da#egg=mmsegmentation
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==2.2
notebook==7.0.2
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.4
nuscenes-devkit==1.1.10
oauthlib==3.2.2
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.16.0
pluggy==1.2.0
plyfile==1.0.1
prettytable==3.8.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.0.9
pyquaternion==0.9.9
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
Shapely==1.8.5
SharedArray==3.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.8.12
tinycss2==1.2.1
tomli==2.0.1
torch==1.10.1+cu113
torchaudio==0.10.1+rocm4.1
torchvision==0.11.2+cu113
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trimesh==2.35.39
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.7
widgetsnbextension==4.0.8
yapf==0.40.1
zipp==3.16.2
